renaming rna_design folder to rnasynth.
