+++ rnasynth/rna_synthesizer.py
@@ -13,9 +13,8 @@ from eden.converter.rna.rnashapes import rnashapes_to_eden
 from eden.util import fit
 from eden.graph import Vectorizer
 
-from rna_design.constraint_extractor import ConstraintExtractor
-# from rna_designer import AntaRNAv109Designer
-from rna_designer import AntaRNAv117Designer
+from rnasynth.constraint_extractor import ConstraintExtractor
+from rnasynth.rna_designer import AntaRNAv117Designer
 
 from util.dataset import binary_classification_dataset_setup
